Moving chat to chat-message.js.
